requirements.txt
openai
langgraph
pydantic
yfinance
backtrader
pandas
numpy
python-dotenv
tinydb
tqdm
ta
matplotlib
plotly
jupyter
